Safari-611.1.10/LayoutTests/svg/animations/svgenum-animation-2-expected.txt

SVG 1.1 dynamic animation tests

Test EdgeModeType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS convolveMatrix.edgeMode.animVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP
PASS convolveMatrix.edgeMode.baseVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP
PASS convolveMatrix.edgeMode.animVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_DUPLICATE
PASS convolveMatrix.edgeMode.baseVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP
PASS convolveMatrix.edgeMode.animVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_DUPLICATE
PASS convolveMatrix.edgeMode.baseVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP
PASS convolveMatrix.edgeMode.animVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_NONE
PASS convolveMatrix.edgeMode.baseVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP
PASS convolveMatrix.edgeMode.animVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_NONE
PASS convolveMatrix.edgeMode.baseVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP
PASS convolveMatrix.edgeMode.animVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP
PASS convolveMatrix.edgeMode.baseVal is SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP
PASS successfullyParsed is true

TEST COMPLETE
